Add support for setting output_wcs of the resample step to an object
